+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/FooInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Symfony\Bundle\FrameworkBundle\Tests\DependencyInjection\Fixtures;
+
+interface FooInterface
+{
+}