+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/WithAsAlias.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Symfony\Bundle\FrameworkBundle\Tests\DependencyInjection\Fixtures;
+
+use Symfony\Component\DependencyInjection\Attribute\AsAlias;
+
+#[AsAlias(id: FooInterface::class)]
+class WithAsAlias
+{
+}