+++ src/Symfony/Component/DependencyInjection/CHANGELOG.md
@@ -18,6 +18,7 @@ CHANGELOG
  * Add support for `ConfigBuilder` in the `PhpFileLoader`
  * Add `ContainerConfigurator::env()` to get the current environment
  * Add `#[Target]` to tell how a dependency is used and hint named autowiring aliases
+ * Add `#[AsAlias]` to tell under which alias a service should be registered
 
 5.2.0
 -----